requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
djangorestframework==3.16.0
playwright==1.52.0
requests==2.32.3
faker==37.3.0
